Added lighting to the game world
Also changed bullets to lasers
gave the lasers a material that should draw them in a certain color (red). The color can be played with by modifying the material.
The laser doesn't have any movement or lifetime, but to give it either one the appropriate lines in update need to be commented out.